Factor out NIF HAL and gpio_helper
This refactors the GPIO code to support pluggable hardware abstraction
layers (HALs) like Circuits.I2C and Circuits.SPI. Currently the options
are sysfs and a stub HAL suitable for testing on non-Linux platforms.

The second change is to pull out the unsafe and Raspberry Pi-specific
code for configuring pull ups and pull downs on GPIOs into a companion
helper app. This app runs in its own process to avoid any risk to
crashing the Erlang VM.